default constructor should return parent value
I came across this issue with a class that caches instances when
they are firste created. Subsequent calls to the constructor will
return the cached instance if available. This relies on a feature
of JavaScript described here:

http://bclary.com/2004/11/07/#a-13.2.2

Note Step 7 in the reference document. By not returning the value
from the default constructor, subclasses of Backbone classes that
rely on this feature will break.