Update backbone.js
In `Events.once` we can streamline the final statements `this.on(...); return this;` into `return this.on(...)`, as `Events.on` also returns `this`. Saves a row for a very slight loss of readability.

Same thing applies in the `_.each` loop where we set up the IoC versions of `on` and `once`.