Added reverse option in the collection sort method
Added the reverse option to be able to sort the collection in the
inverse order specified by the comparator. Useful for one argument
comparators returning string values.

For boolean and integer comparators are easy to find the inverse
function to reverse the collection but for string comparators it's
not as easy as it seems.