Enable ESLint
This pull request replaces the old "jsl" linter with ESLint and enables it as
part of our integration tests.

So far, the only rules it includes are the ones which are both currently
enforced in Underscore AND are already passing in Backbone.

I will follow up with individual pull requests for each rule which will require
changes to the code in an attempt to bring us in sync with Underscore's style.