Run SauceLabs Firefox 11 tests specifically on Windows 10 (#4251)
I tried a couple of different OS version, and this one was the first
where Firefox 11 seems to connect correctly at least some of the time.

(To clarify: the tests were passing in Firefox 11 on all OS versions I
tried, but in most cases a connection error prevented Karma from
knowing this. I only know because I saw the reports in the Sauce web
interface.)